Guia7/Introduccion Java/POO_Ej1/src/Entidad/Libro.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.


ISBN, Título, Autor,
Número de páginas
 */
package Entidad;

/**
 *
 * @author Daniel
 */
public class Libro {
    
    public String isbn;
    public String titulo;
    public String autor;
    public int nPaginas;

    public Libro() {
    }

    public Libro(String isbn, String titulo, String autor, int nPaginas) {
        this.isbn = isbn;
        this.titulo = titulo;
        this.autor = autor;
        this.nPaginas = nPaginas;
    }

    public void setIsbn(String isbn) {
        this.isbn = isbn;
    }

    public void setTitulo(String titulo) {
        this.titulo = titulo;
    }

    public void setAutor(String autor) {
        this.autor = autor;
    }

    public void setnPaginas(int nPaginas) {
        this.nPaginas = nPaginas;
    }

    public String getIsbn() {
        return isbn;
    }

    public String getTitulo() {
        return titulo;
    }

    public String getAutor() {
        return autor;
    }

    public int getnPaginas() {
        return nPaginas;
    }
    
    
    
    
}
